--- src/hooks/selector.tsx
@@ -1,12 +1,0 @@
-import { selector, DefaultValue } from 'recoil';
-import { commentState } from '../states/commentState';
-import { CommentProps } from '../types/comment';
-
-export const addCommentItem = selector({
-    key: 'addCommentItem',
-    get: ({ get }) => get(commentState),
-    set: ({ set }, newItem: CommentProps[] | DefaultValue) => {
-      if (newItem instanceof DefaultValue) return;
-      set(commentState, (prevComment: CommentProps[]) => [...prevComment, ...newItem]);
-    },
-  });
